refactored as its own dpk_similarity named module
