+++ transforms/language/similarity/Dockerfile.python
@@ -9,28 +9,19 @@ RUN pip install --no-cache-dir pytest
 RUN useradd -ms /bin/bash dpk
 USER dpk
 WORKDIR /home/dpk
+
 ARG DPK_WHEEL_FILE_NAME
 
 # Copy and install data processing libraries 
 # These are expected to be placed in the docker context before this is run (see the make image).
-COPY --chown=dpk:root data-processing-dist data-processing-dist
+COPY --chown=dpk:root data-processing-dist/ data-processing-dist/
 RUN  pip install data-processing-dist/${DPK_WHEEL_FILE_NAME}
 
 # END OF STEPS destined for a data-prep-kit base image 
 
-COPY --chown=dpk:root src/ src/
-COPY --chown=dpk:root pyproject.toml pyproject.toml 
-RUN pip install --no-cache-dir -e .
-
-# copy transform main() entry point to the image 
-COPY ./src/similarity_transform_python.py .
-
-# copy some of the samples in
-COPY ./src/similarity_local.py local/
-
-# copy test
-COPY test/ test/
-COPY test-data/ test-data/
+COPY --chown=dpk:root dpk_similarity/ dpk_similarity/
+COPY --chown=dpk:root requirements.txt requirements.txt
+RUN pip install -r requirements.txt
 
 # Set environment
 ENV PYTHONPATH /home/dpk
