--- transforms/language/similarity/python/requirements.txt
@@ -1,6 +1,0 @@
-data-prep-toolkit==0.2.2.dev2
-nltk==3.9.1
-transformers==4.38.2
-torch==2.4.1
-pandas==2.2.2
-requests==2.32.3
