+++ transforms/language/similarity/requirements.txt
@@ -1,0 +1,4 @@
+nltk>=3.9.1
+transformers>=4.38.2
+pandas>=2.2.2
+requests>=2.32.3
